% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRank.R
\name{getRank}
\alias{getRank}
\title{Get higher order taxonomic information for WFO IDs}
\usage{
getRank(
  x,
  rank = NULL,
  useCache = FALSE,
  useAPI = TRUE,
  raw = FALSE,
  capacity = 60,
  fill_time_s = 60,
  timeout = 10,
  wfo_version = NULL
)
}
\arguments{
\item{x}{vector of WFO IDs, e.g. "wfo-0000214110". Note IDs must be of
accepted names. Non-accepted names will not return any higher order
taxonomic information.}

\item{rank}{optional character vector of desired ranks. If NULL (default),
all higher order taxonomic information is returned. Return all available
ranks with \code{listRanks()}}

\item{useCache}{logical, if TRUE use cached values in
\code{wfo_cache_get()} preferentially, to reduce the number of API calls}

\item{useAPI}{logical, if TRUE (default) allow API calls}

\item{capacity}{maximum number of API calls which can accumulate over the
duration of \code{fill_time_s}. See documentation for \code{httr2::req_throttle()}
 @param fill_time_s time in seconds to refill the capacity for repeated API
calls. See documentation for \code{httr2::req_throttle()}}

\item{timeout}{time in seconds to wait before disconnecting from an
unresponsive request}

\item{wfo_version}{optional, WFO backbone release version, as \code{"YYYY-MM"}.
If \code{NULL}, \code{wfoVersion()} provides the most recent version.}
}
\value{
dataframe with higher order taxonomic names of each submitted WFO ID
}
\description{
Get higher order taxonomic information for WFO IDs
}
\examples{
listRanks()
x <- c("wfo-4000011857", "wfo-0000455689")
getRank(x)
getRank(x, raw = TRUE)
getRank(x, rank = c("species", "family")

}
\references{
Borsch, T. et al. (2020). \emph{World Flora Online: Placing taxonomists at the
heart of a definitive and comprehensive global resource on the world's
plants}. TAXON, 69, 6. doi10.1002/tax.12373:
}
\author{
John L. Godlee
}
