% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRank.R
\name{listRanks}
\alias{listRanks}
\title{Return all valid ranks in the World Flora Online database}
\usage{
listRanks()
}
\value{
character vector, e.g. for use in \code{getRank(rank = )}
}
\description{
Return all valid ranks in the World Flora Online database
}
