% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchName.R
\name{matchName}
\alias{matchName}
\title{Correct and match taxonomic names to the World Flora Online database}
\usage{
matchName(
  x,
  interactive = TRUE,
  sub_pattern = subPattern(),
  tolower = TRUE,
  nonumber = TRUE,
  preferAccepted = FALSE,
  preferFuzzy = FALSE,
  useCache = FALSE,
  useAPI = TRUE,
  fallbackToGenus = FALSE,
  checkRank = FALSE,
  checkHomonyms = TRUE,
  fuzzyNameParts = 3,
  raw = FALSE,
  capacity = 60,
  fill_time_s = 60,
  timeout = 10,
  wfo_version = NULL
)
}
\arguments{
\item{x}{vector of taxonomic names}

\item{interactive}{logical, if TRUE (default) user will be prompted to pick
names from a list where multiple ambiguous matches are found, otherwise
names with multiple ambiguous matches will be skipped}

\item{sub_pattern}{character vector of regex patterns which will be removed
from names in \code{x} using \code{gsub()}. The order of this vector matters,
substitutions are applied sequentially. Sensible defaults are provided
by \code{subPattern()}}

\item{tolower}{logical, if TRUE (default) names are converted to all lower case
characters}

\item{nonumber}{logical, if TRUE (default) names containing numbers will be
interpreted as genera, only matching the first word}

\item{preferAccepted}{logical, if TRUE, if multiple ambiguous matches are
found, and if only one candidate is an "accepted" name,
automatically choose that name}

\item{preferFuzzy}{logical, if TRUE, if multiple ambiguous matches are
found, the accepted matched name with the lowest Levenshtein distance to
the submitted name will be returned}

\item{useCache}{logical, if TRUE use cached values in
\code{wfo_cache_get()} preferentially, to reduce the number of API
calls}

\item{useAPI}{logical, if TRUE (default) allow API calls}

\item{fallbackToGenus}{logical, if TRUE genus-level matches will be returned
if no species-level match is available}

\item{checkRank}{logical, if TRUE consider matches to be ambiguous if it is
possible to estimate taxonomic rank from the search string and the rank
does not match that in the name record}

\item{checkHomonyms}{logical, if TRUE (default) consider matches to be
ambiguous if there are other names with the same words but different
author strings}

\item{fuzzyNameParts}{integer value of 3 (default) or greater. The maximum
Levenshtein distance used for server-side fuzzy matching of words in \code{x}
with taxonomic names in the WFO database. This does not affect further
client-side fuzzy matching of multiple ambiguous matches if
\code{preferFuzzy = TRUE}}

\item{raw}{logical, if TRUE raw a nested list is returned, otherwise a
dataframe}

\item{capacity}{maximum number of API calls which can accumulate over the
duration of \code{fill_time_s}. See documentation for \code{httr2::req_throttle()}
 @param fill_time_s time in seconds to refill the capacity for repeated API
calls. See documentation for \code{httr2::req_throttle()}}

\item{timeout}{time in seconds to wait before disconnecting from an
unresponsive request}

\item{wfo_version}{optional, WFO backbone release version, as \code{"YYYY-MM"}.
If \code{NULL}, \code{wfoVersion()} provides the most recent version.}
}
\value{
dataframe containing taxonomic name information with rows matching
names in \code{x}
\describe{
\item{taxon_name_raw}{Original name as in \code{x}}
\item{taxon_name_subm}{Name after optional sanitisation according to
\code{sub_pattern}, \code{tolower}, \code{nonumber}}
\item{method}{The method by which the name was matched. Either: "AUTO" if
a single non-ambiguous accepted name was matched, "AUTO ACC" if
preferAccepted = TRUE and a single accepted name was found among the
possible candidate names, "AUTO FUZZY" if preferFuzzy = TRUE and at least
one accepted name was found among the possible candidate names, "MANUAL"
if interactive = TRUE and the user picked a name, or "EMPTY" if no matches
were found}
\item{fallbackToGenus}{Value of argument in function call}
\item{checkRank}{Value of argument in function call}
\item{checkHomonyms}{Value of argument in function call}
\item{fuzzyNameParts}{Value of argument in function call}
\item{preferAccepted}{Value of argument in function call}
\item{preferFuzzy}{Value of argument in function call}
\item{tolower}{Value of argument in function call}
\item{nonumber}{Value of argument in function call}
\item{taxon_wfo_syn}{WFO ID of matched name}
\item{taxon_name_syn}{Taxonomic name of matched name}
\item{taxon_auth_syn}{Authority of matched name}
\item{taxon_stat_syn}{Taxonomy status of matched name, e.g. "conserved",
"deprecated", "illegitimate", etc}
\item{taxon_role_syn}{Taxonomic role of matched name, e.g. "accepted",
"synonym", "unplaced", etc}
\item{taxon_rank_syn}{Taxonomic rank of matched name, e.g. "species",
"genus", "family", etc}
\item{taxon_wfo_acc}{WFO ID of accepted name}
\item{taxon_name_acc}{Taxonomic name of accepted name}
\item{taxon_auth_acc}{Authority of accepted name}
\item{taxon_stat_acc}{Taxonomy status of accepted name, e.g. "conserved",
"deprecated", "illegitimate", etc}
\item{taxon_role_acc}{Taxonomic role of accepted name, e.g. "accepted",
"synonym", "unplaced", etc}
\item{taxon_rank_acc}{Taxonomic rank of accepted name, e.g. "species",
"genus", "family", etc}
}
}
\description{
This function matches taxonomic names using the World Flora Online database,
via their GraphQL API
}
\examples{
x <- c("Burkea africana", "Julbernardia paniculata", "Fabaceae", 
  "Indet indet", "Brachystegia")
matchName(x)
matchName(x, raw = TRUE)
matchName(x, fallbackToGenus = TRUE)
matchName(x, interactive = FALSE)

}
\references{
Borsch, T. et al. (2020). \emph{World Flora Online: Placing taxonomists at the
heart of a definitive and comprehensive global resource on the world's
plants}. TAXON, 69, 6. doi10.1002/tax.12373:
}
\author{
John L. Godlee
}
