% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchNames.R
\name{matchNames}
\alias{matchNames}
\title{Match a vector of names and return formatted output}
\usage{
matchNames(x, genus_fallback = FALSE, interactive = TRUE, raw = FALSE)
}
\arguments{
\item{x}{vector of taxonomic names
If the column values do not include the author strings for the plant names then a authors_col should be
specified.}

\item{genus_fallback}{logical, if TRUE genus-level matches will be returned
if no species-level match is available}

\item{interactive}{logical, if TRUE (default) user will be prompted to pick
names from a list where multiple ambiguous matches are found, otherwise
names with multiple ambiguous matches will be skipped}

\item{raw}{logical, if TRUE raw a nested list is returned, otherwise a
dataframe}
}
\value{
data.frame containing taxonomic name information with rows matching
names in \code{x}, or a list containing unique values in \code{x}
}
\description{
Match a vector of names and return formatted output
}
\examples{
x <- c("Burkea africana", "Julbernardia paniculata", "Fabaceae", 
  "Indet indet", "Brachystegia")
matchNames(x)
matchNames(x, raw = TRUE)
matchNames(x, genus_fallback = TRUE)
matchNames(x, interactive = FALSE)

}
