% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickName.R
\name{pickName}
\alias{pickName}
\title{Manually pick a taxonomic name from a list returned by the WFO GraphQL API}
\usage{
pickName(x, cand, offset = 0, page_size = 10)
}
\arguments{
\item{x}{original taxonomic name searched by \code{callAPI()}}

\item{cand}{list of candidate taxa returned by \code{callAPI()}}

\item{offset}{index value used internally by pager, controls page length}
}
\value{
list containing information of matched taxonomic name
}
\description{
Manually pick a taxonomic name from a list returned by the WFO GraphQL API
}
\examples{
x <- "Burkea af"
resp <- callAPI(x, query_taxonNameMatch())
pickName(x, resp$data$taxonNameMatch$candidates)

}
