% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subPattern.R
\name{subPattern}
\alias{subPattern}
\title{Common name substitutions}
\usage{
subPattern()
}
\value{
character vector with common name substitutions. Used in \code{gsub()}
}
\description{
Common name substitutions
}
