% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{wfo_cache_load}
\alias{wfo_cache_load}
\title{Load an existing WFO cache object}
\usage{
wfo_cache_load(file = "./wfo_cache.rds")
}
\arguments{
\item{file}{filepath for the cache}
}
\description{
Load an existing WFO cache object
}
