% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{wfo_cache_save}
\alias{wfo_cache_save}
\title{Save a WFO cache to file}
\usage{
wfo_cache_save(file = "./wfo_cache.rds")
}
\arguments{
\item{file}{filepath for the cache}
}
\description{
Save a WFO cache to file
}
