% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfoAPI-package.R
\docType{package}
\name{wfoAPI-package}
\alias{wfoAPI}
\alias{wfoAPI-package}
\title{wfoAPI: Match Taxonomic Names with World Flora Online GraphQL API}
\description{
Search for taxonomic names using the World Flora Online GraphQL API
}
\author{
\strong{Maintainer}: John L. Godlee \email{johngodlee@gmail.com}

}
\keyword{internal}
