% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfoVersion.R
\name{wfoVersion}
\alias{wfoVersion}
\title{Get the most recent WFO taxonomic backbone classification}
\usage{
wfoVersion()
}
\value{
character string with WFO taxonomic backbone classification. A
YYYY-MM date
}
\description{
Get the most recent WFO taxonomic backbone classification
}
