% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getName.R
\name{getName}
\alias{getName}
\title{Get taxonomic information from the World Flora Online database using a WFO ID}
\usage{
getName(
  x,
  useCache = FALSE,
  useAPI = TRUE,
  raw = FALSE,
  capacity = 60,
  fill_time_s = 60,
  timeout = 10
)
}
\arguments{
\item{x}{vector of WFO IDs, e.g. "wfo-0000214110"}

\item{useCache}{logical, if TRUE use cached values in
\code{wfo_cache_get()} preferentially, to reduce the number of API calls}

\item{useAPI}{logical, if TRUE (default) allow API calls}

\item{raw}{logical, if TRUE raw a nested list is returned, otherwise a
dataframe}

\item{capacity}{maximum number of API calls which can accumulate over the
duration of \code{fill_time_s}. See documentation for \code{httr2::req_throttle()}
 @param fill_time_s time in seconds to refill the capacity for repeated API
calls. See documentation for \code{httr2::req_throttle()}}

\item{timeout}{time in seconds to wait before disconnecting from an
unresponsive request}
}
\value{
dataframe with taxonomic name information for each submitted WFO ID
\describe{
\item{taxon_wfo_syn}{WFO ID as in \code{x}}
\item{taxon_name_syn}{Taxonomic name}
\item{taxon_auth_syn}{Authority}
\item{taxon_stat_syn}{Taxonomy status, e.g. "conserved", "deprecated",
"illegitimate", etc}
\item{taxon_role_syn}{Taxonomic role, e.g. "accepted", "synonym",
"unplaced", etc}
\item{taxon_rank_syn}{Taxonomic rank, e.g. "species", "genus",
"family", etc}
\item{taxon_wfo_acc}{WFO ID of accepted name}
\item{taxon_name_acc}{Taxonomic name of accepted name}
\item{taxon_auth_acc}{Authority of accepted name}
\item{taxon_stat_acc}{Taxonomy status of accepted name, e.g. "conserved",
"deprecated", "illegitimate", etc}
\item{taxon_role_acc}{Taxonomic role of accepted name, e.g. "accepted",
"synonym", "unplaced", etc}
\item{taxon_rank_acc}{Taxonomic rank of accepted name, e.g. "species",
"genus", "family", etc}
}
}
\description{
This function retrieves taxonomic information from the World Flora Online
database, via their GraphQL API
}
\examples{
x <- c("wfo-0000187086", "wfo-0000620977", "wfo-4000033298")
getName(x)
getName(x, raw = TRUE)

}
\references{
Borsch, T. et al. (2020). \emph{World Flora Online: Placing taxonomists at the
heart of a definitive and comprehensive global resource on the world's
plants}. TAXON, 69, 6. doi10.1002/tax.12373:
}
\author{
John L. Godlee
}
