% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRank.R
\name{getRank}
\alias{getRank}
\title{Get higher order taxonomic information for WFO IDs}
\usage{
getRank(x, capacity = 60, fill_time_s = 60)
}
\arguments{
\item{x}{vector of WFO IDs, e.g. "wfo-0000214110"}

\item{capacity}{maximum number of API calls which can accumulate over the
duration of \code{fill_time_s}. See documentation for \code{httr2::req_throttle()}
 @param fill_time_s time in seconds to refill the capacity for repeated API
calls. See documentation for \code{httr2::req_throttle()}}
}
\value{
list of dataframes with the higher order taxonomic ranks of each of
the submitted WFO IDs
}
\description{
Get higher order taxonomic information for WFO IDs
}
