% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchName.R
\name{matchName}
\alias{matchName}
\title{Match a taxonomic name against the WFO GraphQL API}
\usage{
matchName(x, genus_fallback = FALSE, interactive = TRUE)
}
\arguments{
\item{x}{taxonomic name to be searched}

\item{genus_fallback}{logical, if TRUE genus-level matches will be returned
if no species-level match is available}

\item{interactive}{logical, if TRUE (default) user will be prompted to pick
names from a list where multiple ambiguous matches are found, otherwise
names with multiple ambiguous matches will be skipped}
}
\value{
list of data returned by API calls
}
\description{
Match a taxonomic name against the WFO GraphQL API
}
\examples{
matchName("Burkea africana")

}
