% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchNames.R
\name{matchNames}
\alias{matchNames}
\title{Correct and match taxonomic names to the World Flora Online database}
\usage{
matchNames(
  x,
  fallbackToGenus = FALSE,
  checkRank = FALSE,
  checkHomonyms = FALSE,
  fuzzyNameParts = 0,
  preferAccepted = FALSE,
  interactive = TRUE,
  useCache = FALSE,
  useAPI = TRUE,
  raw = FALSE,
  delay = 0
)
}
\arguments{
\item{x}{vector of taxonomic names}

\item{fallbackToGenus}{logical, if TRUE genus-level matches will be returned
if no species-level match is available}

\item{checkRank}{logical, if TRUE consider matches to be ambiguous if it is
possible to estimate taxonomic rank from the search string and the rank
does not match that in the name record}

\item{checkHomonyms}{logical, if TRUE consider matches to be ambiguous if
there are other names with the same words but different author strings}

\item{fuzzyNameParts}{integer value of 0 (default) or greater. The maximum
Levenshtein distance used for fuzzy matching words in \code{x}}

\item{preferAccepted}{logical, if TRUE, if multiple ambiguous matches are
found, and if there is only one candidate is an "accepted" name,
automatically choose that name}

\item{interactive}{logical, if TRUE (default) user will be prompted to pick
names from a list where multiple ambiguous matches are found, otherwise
names with multiple ambiguous matches will be skipped}

\item{useCache}{logical, if TRUE use cached values in
\code{.GlobalEnv : wfo_cache} preferentially, to reduce the number of API
calls}

\item{useAPI}{logical, if TRUE (default) allow API calls}

\item{raw}{logical, if TRUE raw a nested list is returned, otherwise a
dataframe}

\item{delay}{number of seconds to pause between API calls. Used to
rate-limit repeated API calls}
}
\value{
data.frame containing taxonomic name information with rows matching
names in \code{x}, or a list containing unique values in \code{x} if raw = TRUE
}
\description{
This function matches taxonomic names using the World Flora Online database,
via their GraphQL API
}
\examples{
x <- c("Burkea africana", "Julbernardia paniculata", "Fabaceae", 
  "Indet indet", "Brachystegia")
matchNames(x)
matchNames(x, raw = TRUE)
matchNames(x, fallbackToGenus = TRUE)
matchNames(x, interactive = FALSE)

}
\references{
Borsch, T. et al. (2020).
\emph{World Flora Online: Placing taxonomists at the heart of a definitive and
comprehensive global resource on the world's plants}. TAXON, 69, 6.
doi10.1002/tax.12373:
}
\author{
John L. Godlee
}
