% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchNames.R
\name{matchNames}
\alias{matchNames}
\title{Correct and match taxonomic names to the World Flora Online database}
\usage{
matchNames(
  x,
  interactive = TRUE,
  sub_pattern = subPattern(),
  tolower = FALSE,
  nonumber = FALSE,
  nobracket = FALSE,
  preferAccepted = FALSE,
  preferFuzzy = FALSE,
  useCache = FALSE,
  useAPI = TRUE,
  fallbackToGenus = FALSE,
  checkRank = FALSE,
  checkHomonyms = TRUE,
  fuzzyNameParts = 3,
  raw = FALSE,
  capacity = 60,
  fill_time_s = 60,
  timeout = 10
)
}
\arguments{
\item{x}{vector of taxonomic names}

\item{interactive}{logical, if TRUE (default) user will be prompted to pick
names from a list where multiple ambiguous matches are found, otherwise
names with multiple ambiguous matches will be skipped}

\item{sub_pattern}{character vector of regex patterns which will be removed
from names in \code{x} using \code{gsub()}. The order of this vector matters,
substitutions are applied sequentially. Sensible defaults are provided
by \code{subPattern()}}

\item{tolower}{logical, if TRUE names are converted to all lower case
characters}

\item{nonumber}{logical, if TRUE names containing numbers will be
interpreted as genera, only matching the first word}

\item{nobracket}{logical If TRUE characters occurring after "(" or ")" will
be removed. Note this could also remove authorities.}

\item{preferAccepted}{logical, if TRUE, if multiple ambiguous matches are
found, and if only one candidate is an "accepted" name,
automatically choose that name}

\item{preferFuzzy}{logical, if TRUE, if multiple ambiguous matches are
found, the accepted matched name with the lowest Levenshtein distance to
the submitted name will be returned}

\item{useCache}{logical, if TRUE use cached values in
\code{wfo_cache_get()} preferentially, to reduce the number of API
calls}

\item{useAPI}{logical, if TRUE (default) allow API calls}

\item{fallbackToGenus}{logical, if TRUE genus-level matches will be returned
if no species-level match is available}

\item{checkRank}{logical, if TRUE consider matches to be ambiguous if it is
possible to estimate taxonomic rank from the search string and the rank
does not match that in the name record}

\item{checkHomonyms}{logical, if TRUE (default) consider matches to be
ambiguous if there are other names with the same words but different
author strings}

\item{fuzzyNameParts}{integer value of 3 (default) or greater. The maximum
Levenshtein distance used for server-side fuzzy matching of words in \code{x}
with taxonomic names in the WFO database. This does not affect further
client-side fuzzy matching of multiple ambiguous matches if
\code{preferFuzzy = TRUE}}

\item{raw}{logical, if TRUE raw a nested list is returned, otherwise a
dataframe}

\item{capacity}{maximum number of API calls which can accumulate over the
duration of \code{fill_time_s}. See documentation for \code{httr2::req_throttle()}
 @param fill_time_s time in seconds to refill the capacity for repeated API
calls. See documentation for \code{httr2::req_throttle()}}

\item{timeout}{time in seconds to wait before disconnecting from an
unresponsive request}
}
\value{
data.frame containing taxonomic name information with rows matching
names in \code{x}, or a list containing unique values in \code{x} if raw = TRUE
}
\description{
This function matches taxonomic names using the World Flora Online database,
via their GraphQL API
}
\examples{
x <- c("Burkea africana", "Julbernardia paniculata", "Fabaceae", 
  "Indet indet", "Brachystegia")
matchNames(x)
matchNames(x, raw = TRUE)
matchNames(x, fallbackToGenus = TRUE)
matchNames(x, interactive = FALSE)

}
\references{
Borsch, T. et al. (2020).
\emph{World Flora Online: Placing taxonomists at the heart of a definitive and
comprehensive global resource on the world's plants}. TAXON, 69, 6.
doi10.1002/tax.12373:
}
\author{
John L. Godlee
}
